154171 - BESS builder service code upliftment  (#163)
* 140032 - code refactored AB#140032

---------

+++ UKHO.PeriodicOutputService/UKHO.PeriodicOutputService.Common/Helpers/IAzureBlobStorageService.cs
@@ -4,5 +4,5 @@ namespace UKHO.PeriodicOutputService.Common.Helpers;
 
 public interface IAzureBlobStorageService
 {
-    Task<bool> SetConfigQueueMessageModelAndAddToQueueAsync(BessConfig bessConfig, IEnumerable<string> encCellNames, int? fileSize);
+    Task<bool> SetConfigQueueMessageModelAndAddToQueueAsync(BessConfig bessConfig, IEnumerable<string> encCellNames, long? fileSize);
 }
